Support data inference schema change function
