+++ inlong-sort/sort-flink/base/src/main/java/org/apache/inlong/sort/base/sink/MultipleSinkOption.java
@@ -17,6 +17,8 @@
 
 package org.apache.inlong.sort.base.sink;
 
+import org.apache.inlong.sort.schema.TableChange;
+
 import org.apache.flink.shaded.guava18.com.google.common.collect.ImmutableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
